# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cms/media:
    get:
      operationId: listCmsMedia
      tags:
        - CMS Media
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Media list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaListResponse'
    post:
      operationId: registerCmsMedia
      tags:
        - CMS Media
      summary: Register an uploaded media object (after presigned upload)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterMediaBody'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaItem'
        '400':
          $ref: '#/components/responses/BadRequest'
  /cms/media/{id}:
    delete:
      operationId: deleteCmsMedia
      tags:
        - CMS Media
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /cms/comments:
    get:
      operationId: listCmsComments
      tags:
        - CMS Comments
      parameters:
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/CommentStatus'
        - in: query
          name: postId
          schema:
            type: string
            format: uuid
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Comments page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentListResponse'
  /cms/comments/{id}/moderate:
    post:
      operationId: moderateCmsComment
      tags:
        - CMS Comments
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - action
              properties:
                action:
                  type: string
                  enum:
                    - approve
                    - reject
                    - spam
                    - delete
      responses:
        '200':
          description: Moderated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
...
components:
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
...
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          type: string
        details: {}
...
    CommentStatus:
      type: string
      enum:
        - pending
        - approved
        - spam
        - deleted
...
    MediaItem:
      type: object
      required:
        - id
        - storageKey
        - publicUrl
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        storageKey:
          type: string
        publicUrl:
          type: string
        mime:
          type: string
          nullable: true
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true
        byteSize:
          type: integer
          nullable: true
        altText:
          type: string
          nullable: true
        uploadedBy:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
    MediaListResponse:
      type: object
      required:
        - items
        - page
        - pageSize
        - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/MediaItem'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
    RegisterMediaBody:
      type: object
      required:
        - storageKey
        - publicUrl
      properties:
        storageKey:
          type: string
          minLength: 1
        publicUrl:
          type: string
          minLength: 1
        mime:
          type: string
          nullable: true
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true
        byteSize:
          type: integer
          nullable: true
        altText:
          type: string
          nullable: true
...
    Comment:
      type: object
      required:
        - id
        - postId
        - authorName
        - authorEmail
        - bodyText
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        postId:
          type: string
          format: uuid
        parentCommentId:
          type: string
          format: uuid
          nullable: true
        authorName:
          type: string
        authorEmail:
          type: string
        bodyText:
          type: string
        status:
          $ref: '#/components/schemas/CommentStatus'
        createdAt:
          type: string
          format: date-time
        moderatedAt:
          type: string
          format: date-time
          nullable: true
    CommentListResponse:
      type: object
      required:
        - items
        - page
        - pageSize
        - total
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
...