# lib/api-zod/src/generated/types/commentStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type CommentStatus = (typeof CommentStatus)[keyof typeof CommentStatus];

export const CommentStatus = {
  pending: "pending",
  approved: "approved",
  spam: "spam",
  deleted: "deleted",
} as const;