# lib/api-zod/src/generated/types/mediaListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MediaItem } from "./mediaItem";

export interface MediaListResponse {
  items: MediaItem[];
  page: number;
  pageSize: number;
  total: number;
}